[#162795470] Form fields accept prop adapter function
The intent of this is to accept any type of prop but only use the ones we accept, for that a prop adapter function is provided where whatever props passed to the component will be turned into the accepted props by the component. This sounds weird but in order to make the field work with certain form hadlers such as redux form, it is necessary for us to allow the component to adapt itself by accepting the adapter.