[#166356555] Commit .nvmrc file into version system again.
There was not a clear explanation why that file removed from Git. The version
included in `.nvmrc` has a different purpose than the version included in
`package.json`. The main goal of the `.nvmrc` is that all developers use `nvm`
to install the same node version in their computers for `aemon` development. The
version included in `.nvmrc` is `10.14.2` which is the latest LTS version that
complies with the one defined in `pakage.json`.